services/data_connectors/pyproject.toml
[project]
name = "data_connectors"
version = "0.1.0"
dependencies = [
    "requests",
    "pydantic",
    "google-api-python-client>=2.185.0",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "lib_infisical",
    "dotenv>=0.9.9",
]

[uv.tool.sources]
lib_infisical = { workspace = true }

[dependency-groups]
dev = [
    "ruff",
    "black",
    "mypy",
    "pytest",
    "pytest-cov",
]
